+++ stylua-vscode/CHANGELOG.md
@@ -20,6 +20,8 @@ To view the changelog of the StyLua binary, see [here](https://github.com/Johnny
 - Added configuration option `stylua.configPath` to provide a direct path to a `stylua.toml` file. Note: this will override any workspace config lookup
 - Added configuration option `stylua.verify` to pass `--verify` to StyLua CLI when formatting a file. This enforces output verification
 - Added command `StyLua: Select Version` to customize which version of StyLua to install. This command updates the `stylua.targetReleaseVersion` setting
+- Added a new language status bar item to display StyLua information
+- Current StyLua version will now be shown in the status bar item
 
 ### Changed